フォーム実装をコンポーネントに変更
