+++ pages/index.vue
@@ -2,18 +2,7 @@
   <div class="container">
     <div>
       <Logo />
-      <h1 class="title">
-        ToDo List
-      </h1>
-      <h2 class="subtitle">
-        タスクを追加してください
-      </h2>
-      <div>
-        <input v-model="content" type="text" />
-        <button @click="addTodo">
-          追加
-        </button>
-      </div>
+      <todo-form />
       <div class="todo-list" align="center">
         <div>
           <input v-model="taskState" type="radio" value="allState" />全て
@@ -54,14 +43,15 @@
 
 <script>
 import Logo from '~/components/Logo.vue'
+import TodoForm from '~/components/TodoForm.vue'
 
 export default {
   components: {
-    Logo
+    Logo,
+    TodoForm
   },
   data() {
     return {
-      content: '',
       taskState: 'allState'
     }
   },
@@ -88,12 +78,6 @@ export default {
     await store.dispatch('todo/fetchTodos')
   },
   methods: {
-    // タスクの追加
-    async addTodo() {
-      await this.$store.dispatch('todo/addTodo', this.content)
-      this.content = ''
-      console.log('addTodo', this.$store.state.todo.todos)
-    },
     // stateボタンの状態切り替え
     updateState(index) {
       this.$store.dispatch('todo/updateState', index)
@@ -107,31 +91,12 @@ export default {
 }
 </script>
 
-<style>
+<style scoped>
 .container {
   margin: 0 auto;
   min-height: 50vh;
-  display: flex;
   justify-content: center;
   align-items: center;
   text-align: center;
-}
-
-.title {
-  font-family: 'Quicksand', 'Source Sans Pro', -apple-system, BlinkMacSystemFont,
-    'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif;
-  display: block;
-  font-weight: 300;
-  font-size: 100px;
-  color: #35495e;
-  letter-spacing: 1px;
-}
-
-.subtitle {
-  font-weight: 300;
-  font-size: 42px;
-  color: #526488;
-  word-spacing: 5px;
-  padding-bottom: 15px;
 }
 </style>
